+++ CHANGES.rst
@@ -5,10 +5,10 @@
 1.1.0 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Fix error on export if no categories where selected in master data.
 
 
 1.0.0 (2015-05-24)
 ==================
 
-* Initial release.
+- Initial release.